# Speed up debug logging, when filtered out